MicroServices/ModelAssemblyService/requirements.txt
FMPy==0.3.5
lxml==4.8.0
SPARQLWrapper==2.0.0
paho-mqtt==1.6.1
